% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgconfig.R
\name{pkgconfig}
\alias{pkgconfig}
\title{Configure compilation flags}
\usage{
pkgconfig(opt = c("PKG_CPPFLAGS", "PKG_LIBS"))
}
\arguments{
\item{opt}{String specifying the \code{Makevars} option to print.}
}
\value{
Flags to add to the requested \code{opt} of the \code{Makevars} are printed to the screen.
}
\description{
Configure flags for compiling downstream packages.
}
\examples{
pkgconfig("PKG_CPPFLAGS")
pkgconfig("PKG_LIBS")

}
\author{
Aaron Lun
}
